[update]DBのテーブルを最新にアップデート
Close #8

テーブルを全体的に治すと同時に、CLIでテーブルを一気に作成できるようにしました。
