+++ sksk_app/__init__.py
@@ -12,11 +12,11 @@ def create_app():
 
     db.init_app(app) 
 
-    # from sksk_app.script.db import qtdb
+    from sksk_app.script.database import qtdb
     from sksk_app.views.pages import pg
     from sksk_app.views.auth import auth
     from sksk_app.views.questions import qt
-    # app.register_blueprint(qtdb)
+    app.register_blueprint(qtdb)
     app.register_blueprint(pg)
     app.register_blueprint(auth)
     app.register_blueprint(qt)
