--- sksk_app/script/db.py
@@ -1,16 +1,0 @@
-# from flask.cli import with_appcontext
-# from sksk_app import create_app
-# from sksk_app.models import db
-# from flask import current_app
-# from flask import Blueprint
-
-# qtdb = Blueprint('qtdb', __name__)
-
-# コマンド機能の作成
-# @qtdb.cli.command(name='create')
-# @with_appcontext
-# def create():
-    # print("コマンド成功！")
-# コマンド機能をcli commandsに追加
-
-# current_app.cli.add_command(create)
